+++ src/index.ts
@@ -3,21 +3,19 @@
 import Task from './task/Task';
 
 /**
- * Hypergiant is used to create signal-like events that run when the Hypergiant
- * is emitted.
- * 
- * One of the big advtanges that Hypergiant has over native JavaScript events is
- * the ability of using named events instead of relying on typing.
- * 
+ * Hypergiant is used to create signals that run a task when emitted.
+ *
+ * One of the biggest advtantages that signals have over native JavaScript events
+ * is that they don't rely on correct typing.
+ *
  * @author Robert Corponoi
  * 
- * @version 0.1.0
+ * @version 2.3.0
  */
 export default class Hypergiant {
 
-	/**
-	 * The tasks associated with this instance of Hypergiant that run
-	 * when the event is emitted.
+  /**
+   * The tasks that are set to run when the corresponding signal is dispatched.
 	 * 
 	 * @since 0.1.0
 	 * 
@@ -26,12 +24,12 @@ export default class Hypergiant {
 	private tasks: Set<Task> = new Set();
 
 	/**
-	 * Adds a new task for this instance of Hypergiant to run when emitted.
+	 * Add a new signal.
 	 * 
 	 * @since 0.1.0
 	 * 
-	 * @param {Function} fn The method that should be called when emitted.
-	 * @param {boolean} [once=false] Indicates whether this task should only be called once and then deleted.
+	 * @param {Function} fn The method that should be called when the signal is dispatched.
+	 * @param {boolean} [once=false] Indicates whether this signal should only be dispatched once and then deleted.
 	 * 
 	 * @returns {Hypergiant} Returns this for chaining.
 	 */
@@ -61,6 +59,53 @@ export default class Hypergiant {
 
 		}
 
-	}
+  }
+
+  /**
+   * Removes a task from this signal by name.
+   *
+   * @since 2.3.0
+   *
+   * @param {Function} The task to remove.
+   *
+   * @returns {Hypergiant} Returns this for chaining.
+   */
+  remove(fn: Function): Hypergiant {
+
+    const fnToString: string = fn.toString();
+
+    for (const task of this.tasks) {
+
+      const taskFnToString: string = task.fn.toString();
+
+      if (fnToString === taskFnToString) {
+
+        this.tasks.delete(task);
+
+        break;
+
+      }
+
+    }
+
+    return this;
+
+  }
+
+  /**
+   * Removes all tasks from this signal.
+   *
+   * @since 2.3.0
+   *
+   * @returns {Hypergiant} Returns this for chaining.
+   */
+  removeAll(): Hypergiant {
+
+    this.tasks.clear();
+
+    return this;
+
+
+  }
 
 }